[QMS-1128] Dev: extract CMapVRT/CDemVRT's duplicated GDAL helpers to CGdalVrtUtil
allReferencedFilesExist(), collectOverviewFactors(), closeDataset() and
progressCallback() were kept as separate, near-identical copies in CMapVRT
and CDemVRT. They operate purely on GDAL types (or, for progressCallback(),
on IDrawContext - the base both CMapDraw and CDemDraw already implement
needsRedraw() through unchanged), with no map- or DEM-specific logic, so
there's no scenario where the two copies should ever behave differently -
only one where someone fixes a bug in one and forgets the other. That had
already started happening: allReferencedFilesExist()'s loop variable had
drifted to int in one copy and qint32 in the other.

Move all four into helpers/CGdalVrtUtil, used by both classes. progressCallback()
collapses from two near-identical functions into one, since it only ever needed
the shared IDrawContext interface, not the concrete CDemDraw/CMapDraw subclass.

Left the constructors' warp-setup blocks and draw() duplicated despite looking
similar: the shared-looking parts are interleaved with genuinely different
decisions (band/color-table validation, resample algorithm choice, alpha-band
synthesis, outline-polygon vs bounding-box derivation, single-image composite
vs parallel multi-layer shading), so unifying them would mean threading several
parameters/callbacks through a shared function for little benefit.
