+++ src/qmapshack/CMakeLists.txt
@@ -182,6 +182,7 @@ set( SRCS
     helpers/CDraw.cpp
     helpers/CElevationDialog.cpp
     gis/search/CSearch.cpp
+    helpers/CGdalVrtUtil.cpp
     helpers/CInputDialog.cpp
     helpers/CLimit.cpp
     helpers/CLinksDialog.cpp
@@ -554,6 +555,7 @@ set( HDRS
     helpers/CDraw.h
     helpers/CElevationDialog.h
     helpers/CFileExt.h
+    helpers/CGdalVrtUtil.h
     gis/search/CSearch.h
     helpers/CInputDialog.h
     helpers/CLimit.h
