+++ src/qmapshack/dem/CDemVRT.h
@@ -27,7 +27,6 @@
 
 class CDemDraw;
 class GDALDataset;
-class GDALRasterBand;
 
 /**
    @brief GDAL-backed IDem implementation: reads elevation data from any raster format
@@ -121,52 +120,6 @@ class CDemVRT : public IDem {
      @return false if pos lies outside the DEM's bounding box
    */
   bool toRasterPixel(const QPointF& pos, QPointF& pixel) const;
-
-  /**
-     @brief Check that every file GDAL reports as part of the dataset (e.g. the files a
-            VRT references) actually exists on disk.
-     @param dataset     the dataset to check
-     @param missingFile set to the first referenced file that could not be found; left
-                         unchanged if all files exist
-     @return false if a referenced file is missing
-   */
-  static bool allReferencedFilesExist(GDALDataset* dataset, QString& missingFile);
-
-  /**
-     @brief Collect virtual-overview decimation factors for dataset, in dataset's own pixel
-            scale.
-
-     Prefers dataset's own overview list. If dataset reports none - e.g. a gdalbuildvrt
-     mosaic whose <OverviewList> is stale, was never written because not every source had
-     overviews when the mosaic was built, or still has one source that lacks them - falls
-     back to the union of overview factors found across the individual files dataset depends
-     on, so one under-prepared source no longer disables overview-accelerated reads for the
-     whole mosaic. Each file's factors are converted via its own geotransform pixel size
-     rather than reused as raw pixel ratios, so sources of differing native resolution stay
-     consistent with dataset's own grid.
-     @param dataset    the (pre-warp) dataset to collect overview factors for
-     @param pBand      dataset's band 1
-     @param pixelSizeX the real-world size of one of dataset's own pixels along x; pass 0 if
-                       unknown to skip the per-file fallback entirely
-     @return sorted, de-duplicated decimation factors; empty if none are available anywhere
-   */
-  static QVector<qint32> collectOverviewFactors(GDALDataset* dataset, GDALRasterBand* pBand, qreal pixelSizeX);
-
-  /**
-     @brief GDAL progress callback aborting the read once a newer redraw has been
-            requested.
-     @param pProgressArg the CDemDraw passed in as progress callback context
-   */
-  static int progressCallback(double dfComplete, const char* message, void* pProgressArg);
-
-  /**
-     @brief Close a GDAL dataset and reset the pointer, tolerating a null dataset.
-
-     GDALClose() logs a CPL error when passed a null handle instead of just ignoring
-     it, so every call site needs this guard - which happens whenever no warped VRT
-     was needed (srcDataset stays null) or construction failed before dataset was set.
-   */
-  static void closeDataset(GDALDataset*& dataset);
 
   /// guards all access to dataset/srcDataset (see class-level @note)
   mutable QMutex mutex;
